requirements.txt
accelerate==0.26.1
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
asgiref==3.7.2
attrs==23.2.0
azure-core==1.29.7
azure-identity==1.15.0
azure-storage-queue==12.9.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blis==0.7.11
bs4==0.0.2
build==1.0.3
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.23
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.4
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.3
datasets==2.16.1
Deprecated==1.2.14
dill==0.3.7
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
docx2txt==0.8
elastic-transport==8.12.0
elasticsearch==8.12.0
evaluate==0.4.1
fastapi==0.109.0
fastapi-injector==0.5.3
ffmpeg==1.4
ffmpy==0.3.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.28.1
googleapis-common-protos==1.62.0
gptcache==0.1.43
gradio==4.16.0
gradio_client==0.8.1
greenlet==3.0.3
grpcio==1.62.0
guidance==0.0.64
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
injector==0.21.0
install==1.3.5
interegular==0.3.3
isodate==0.6.1
Jinja2==3.1.3
joblib==1.3.2
jsonpath-ng==1.6.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==29.0.0
langcodes==3.3.0
llama-index==0.10.12
llama-index-agent-openai==0.1.5
llama-index-cli==0.1.5
llama-index-core==0.10.12
llama-index-embeddings-azure-openai==0.1.6
llama-index-embeddings-huggingface==0.1.3
llama-index-embeddings-openai==0.1.6
llama-index-indices-managed-llama-cloud==0.1.3
llama-index-legacy==0.9.48
llama-index-llms-azure-openai==0.1.4
llama-index-llms-openai==0.1.6
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.4
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.5
llama-index-readers-llama-parse==0.1.3
llama-index-vector-stores-chroma==0.1.4
llama-index-vector-stores-elasticsearch==0.1.3
llama-parse==0.3.4
llamaindex-py-client==0.1.13
lm-format-enforcer==0.4.3
loguru==0.7.2
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.2
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.26.0
msal-extensions==1.1.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.9
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.17.0
openai==1.8.0
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
optimum==1.16.1
orjson==3.9.12
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
platformdirs==4.1.0
ply==3.11
portalocker==2.8.2
posthog==3.4.2
preshed==3.0.9
protobuf==4.25.2
psutil==5.9.7
pulsar-client==3.4.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
pygtrie==2.5.0
PyJWT==2.8.0
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
pyparsing==3.1.1
pypdf==4.0.2
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
python-pptx==0.6.23
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
rank-bm25==0.2.2
redis==5.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruff==0.1.15
safetensors==0.4.1
scikit-learn==1.4.0
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.40.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
srsly==2.4.8
starlette==0.35.1
sympy==1.12
tenacity==8.2.3
thinc==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.0
torchaudio==2.2.0
tqdm==4.66.1
transformers==4.37.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.26.0
wandb==0.16.2
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websocket-client==1.7.0
websockets==11.0.3
win32-setctime==1.1.0
wrapt==1.16.0
XlsxWriter==3.1.9
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
